Investigation of string-based topology finding framework for structured surface patterns
This research aims to address novel implementation of learning algorithms and generative design to stringbased topology exploration methods. It seeks to achieve flexibility in the generation of high-performing
diverse structural patterns of shell-like structures that effectively align multiple engineering objectives
including mechanical and material efficiency and buildability. Through an investigation into the tailored
application of reinforcement learning and genetic algorithms for connectivity design within structural
patterns, both qualitative and quantitative metrics will be defined to demonstrate the strength and
generality of this approach. Ultimately, the research aims to facilitate general creative exploration in
structural design during the conceptual stages, emphasizing the importance of closer collaboration
between form-designers and form-analyzers to harness the potential of emerging computation techniques.